Configure Devise Confirmable via Email
In this commit, we modify our initial devise_create_users migration to
add an index to the username field, then we create a new migration
called add_confirmable_to_devise. We add the required fields for a
confirmable devise user.

We update the user model to make sure that:
• First Name is required
• Last Name is required
• Username is required and must be unique

Next, we modify the development.rb file to include smtp_settings for
action_mailer.

Next, we update the secrets.yml file to include environment variables
used in the smtp_settings configuration, including:
• domain_name
• email_provider_username
• email_provider_password

Finally, we update the application controller to include a before_filter
that configures our custom devise parameters.